- Forgot to add dists.dss file used in dbgen.
- Minor fixes (e.g., _scaling_factor in TPC-H)

- TPC-H SF=0.1 (~215MB) takes ~405secs to load with -g

+++ include/sm/shore/shore_env.h
@@ -448,10 +448,11 @@ protected:
     // TM1:  SF=1 --> 15B   (10K Subscribers)
     // TPCB: SF=1 --> 20MB  (1 Branch)
     // TPCC: SF=1 --> 130MB (1 Warehouse)
+    // TPCH: SF=1 --> 250MB (1 Warehouse)
     //       
-    int             _scaling_factor; 
+    double             _scaling_factor; 
     pthread_mutex_t _scaling_mutex;
-    int             _queried_factor;
+    double             _queried_factor;
     pthread_mutex_t _queried_mutex;
